Fixed issue crashing an app in case of absence file

+++ data-crawler/src/main/java/util/PdfDownloader.java
@@ -9,23 +9,46 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 public class PdfDownloader {
-    public static FileInputStream downloadAndGetAsStream(String url) throws DataCrawlerException {
-        try {
-            URL link = new URL(url);
-            String tempDirPath = FileUtils.getTempDirectoryPath();
-            String pdfDir = tempDirPath + "medicalPdf";
-            FileUtils.forceMkdir(new File(pdfDir));
-            String pdfName = FilenameUtils.getBaseName(url) + "." + FilenameUtils.getExtension(url);
-            String pathToPdfFile = pdfDir + File.separator + pdfName;
-            File pdfFile = new File(pathToPdfFile);
-            FileUtils.copyURLToFile(link, pdfFile);
-            return FileUtils.openInputStream(pdfFile);
-        } catch (MalformedURLException e) {
-            throw new DataCrawlerException("Wrong url: " + url, e);
-        } catch (IOException e) {
-            throw new DataCrawlerException("Some IO error", e);
-        }
-    }
+	private static String BASE_PATH;
+
+	static {
+		String tempDirPath = FileUtils.getTempDirectoryPath();
+		BASE_PATH = tempDirPath + "medicalPdf";
+
+	}
+
+	public static FileInputStream downloadAndGetAsStream(String url) throws DataCrawlerException {
+		try {
+			File baseDir = new File(BASE_PATH);
+			if (!baseDir.exists()) {
+				FileUtils.forceMkdir(new File(BASE_PATH));
+			}
+
+			String pdfName = FilenameUtils.getBaseName(url) + "." + FilenameUtils.getExtension(url);
+			String pathToPdfFile = BASE_PATH + File.separator + pdfName;
+
+			Path path = Paths.get(pathToPdfFile);
+
+			if (Files.notExists(path)) {
+				return null;
+			}
+
+			File pdfFile = new File(pathToPdfFile);
+
+			URL link = new URL(url);
+			FileUtils.copyURLToFile(link, pdfFile);
+
+			return FileUtils.openInputStream(pdfFile);
+
+		} catch (MalformedURLException e) {
+			throw new DataCrawlerException("Wrong url: " + url, e);
+		} catch (IOException e) {
+			throw new DataCrawlerException("Some IO error", e);
+		}
+	}
 }
